LofiHollow: NPCs
Implemented and finished almost everything having to do with NPCs, including the dialogue system, giving items to them, and buying/selling items from shops.

Streamlined maps so that only the needed ones are loaded. Fixed the minimap to work with the new method.

Made monster definitions able to have defined templates that give them a set list of bonuses, strengths, and weaknesses.